add discog helper and refactor methods
