+++ Gemfile
@@ -21,6 +21,7 @@ end
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
+  gem 'awesome_print'
   gem 'pry-rails'
   gem 'pry-byebug'
   gem 'rspec-rails'
